libibverbs: Split Linux-specific initialization into dedicated source
