+++ libibverbs/ibverbs.h
@@ -90,6 +90,7 @@ static inline const struct verbs_context_ops *get_ops(struct ibv_context *ctx)
 enum ibv_node_type decode_knode_type(unsigned int knode_type);
 
 int find_sysfs_devs_nl(struct list_head *tmp_sysfs_dev_list);
+int find_sysfs_devs(struct list_head *tmp_sysfs_dev_list);
 
 int try_access_device(const struct verbs_sysfs_dev *sysfs_dev);
